+++ weaver/processes/wps_package.py
@@ -104,6 +104,7 @@ from weaver.vault.utils import (
     map_vault_location,
     parse_vault_token
 )
+from weaver.wps.storage import ReferenceStatusLocationStorage
 from weaver.wps.utils import get_wps_output_dir, get_wps_output_url, map_wps_output_location
 from weaver.wps_restapi import swagger_definitions as sd
 
@@ -184,6 +185,8 @@ def retrieve_package_job_log(execution, job, progress_min=0, progress_max=100):
     try:
         # weaver package log every status update into this file (we no longer rely on the http monitoring)
         out_dir = get_wps_output_dir(get_settings())
+        if job.context:
+            out_dir = os.path.join(out_dir, job.context)
         # if the process is a weaver package this status xml should be available in the process output dir
         log_path = get_status_location_log_path(execution.statusLocation, out_dir=out_dir)
         with open(log_path, mode="r", encoding="utf-8") as log_file:
@@ -904,6 +907,7 @@ class WpsPackage(Process):
         self.request = None                  # type: Optional[WorkerRequest]
         self.response = None                 # type: Optional[ExecuteResponse]
         self._job = None                     # type: Optional[Job]
+        self._job_status_file = None         # type: Optional[str]
 
         self.payload = payload
         self.package = package
@@ -931,6 +935,30 @@ class WpsPackage(Process):
             status_supported=True,
             **kw
         )
+
+    @property
+    def status_filename(self):
+        # type: () -> str
+        """
+        Obtain the XML status location of this process when executed.
+
+        The status location applies the ``WPS-Output-Context`` if defined such that any following output
+        or log file references that derive from it will be automatically stored in the same nested context.
+        """
+        if self._job_status_file:
+            return self._job_status_file
+        # avoid error in case process execution is a remote provider (eg: during workflow step)
+        # file is already defined in a specific location and cannot be moved
+        if isinstance(self.status_store, ReferenceStatusLocationStorage):
+            self._job_status_file = self.status_store.location()
+            return self._job_status_file
+        status_file = super(WpsPackage, self).status_filename
+        if self.job.context:
+            status_store_ctx_dir = os.path.join(self.status_store.target, self.job.context)
+            os.makedirs(status_store_ctx_dir, exist_ok=True)
+            status_file = os.path.join(self.job.context, status_file)
+        self._job_status_file = status_file
+        return status_file
 
     def setup_loggers(self, log_stdout_stderr=True):
         # type: (bool) -> None